Extract AgentBehavior operations into document_config/behavior.rs
