+++ crates/defra-agent/src/document_config/mod.rs
@@ -8,29 +8,20 @@ mod graphql_fields;
 pub use principal::{load_agent_principal, upsert_agent_principal, AgentPrincipal};
 pub(crate) use principal::{load_agent_principal_by_doc_id, load_agent_principal_record};
 
+pub use behavior::{
+    list_agent_behaviors, load_agent_behavior, upsert_agent_behavior, AgentBehavior,
+};
+#[allow(unused_imports)]
+pub(crate) use behavior::{
+    list_agent_behavior_records, load_agent_behavior_by_doc_id, load_agent_behavior_record,
+};
+use behavior::create_default_behavior;
+
 use anyhow::{anyhow, Result};
 use defra_node::EmbeddedNode;
 use serde::{Deserialize, Serialize};
 
 use crate::graphql::escape_graphql_string;
-
-const DEFAULT_BEHAVIOR_LABEL: &str = "Default";
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct AgentBehavior {
-    pub behavior_id: String,
-    pub agent_did: String,
-    pub display_name: Option<String>,
-    pub system_prompt: Option<String>,
-    pub backend_id: Option<String>,
-    pub model_name: Option<String>,
-    pub tool_selection_id: Option<String>,
-    pub inference_profile_id: Option<String>,
-    pub compaction_strategy: Option<String>,
-    pub compaction_threshold: Option<f64>,
-    pub enabled: bool,
-    pub created_at: Option<String>,
-}
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct InferenceProfile {
@@ -166,134 +157,6 @@ pub async fn ensure_agent_principal(
         created_principal,
         created_default_behavior,
     })
-}
-
-pub async fn load_agent_behavior(
-    node: &EmbeddedNode,
-    behavior_id: &str,
-) -> Result<Option<AgentBehavior>> {
-    Ok(load_agent_behavior_record(node, behavior_id)
-        .await?
-        .map(|(_, behavior)| behavior))
-}
-
-pub(crate) async fn load_agent_behavior_record(
-    node: &EmbeddedNode,
-    behavior_id: &str,
-) -> Result<Option<(String, AgentBehavior)>> {
-    let escaped_behavior_id = escape_graphql_string(behavior_id);
-    let query = format!(
-        r#"{{
-            AgentBehavior(
-                filter: {{ behavior_id: {{ _eq: "{escaped_behavior_id}" }} }},
-                limit: 1
-            ) {{
-                _docID
-                behavior_id
-                agent_did
-                display_name
-                system_prompt
-                backend_id
-                model_name
-                tool_selection_id
-                inference_profile_id
-                compaction_strategy
-                compaction_threshold
-                enabled
-                created_at
-            }}
-        }}"#
-    );
-
-    let resp = node.execute(&query).await;
-    if resp.has_errors() {
-        anyhow::bail!("query AgentBehavior failed: {:?}", resp.errors);
-    }
-
-    Ok(serde_helpers::first_row_with_doc_id(resp.data.as_ref(), "AgentBehavior"))
-}
-
-pub(crate) async fn load_agent_behavior_by_doc_id(
-    node: &EmbeddedNode,
-    doc_id: &str,
-) -> Result<Option<(String, AgentBehavior)>> {
-    let escaped_doc_id = escape_graphql_string(doc_id);
-    let query = format!(
-        r#"{{
-            AgentBehavior(
-                filter: {{ _docID: {{ _eq: "{escaped_doc_id}" }} }},
-                limit: 1
-            ) {{
-                _docID
-                behavior_id
-                agent_did
-                display_name
-                system_prompt
-                backend_id
-                model_name
-                tool_selection_id
-                inference_profile_id
-                compaction_strategy
-                compaction_threshold
-                enabled
-                created_at
-            }}
-        }}"#
-    );
-
-    let resp = node.execute(&query).await;
-    if resp.has_errors() {
-        anyhow::bail!("query AgentBehavior by _docID failed: {:?}", resp.errors);
-    }
-
-    Ok(serde_helpers::first_row_with_doc_id(resp.data.as_ref(), "AgentBehavior"))
-}
-
-pub async fn list_agent_behaviors(
-    node: &EmbeddedNode,
-    agent_did: &str,
-) -> Result<Vec<AgentBehavior>> {
-    Ok(list_agent_behavior_records(node, agent_did)
-        .await?
-        .into_iter()
-        .map(|(_, behavior)| behavior)
-        .collect())
-}
-
-pub(crate) async fn list_agent_behavior_records(
-    node: &EmbeddedNode,
-    agent_did: &str,
-) -> Result<Vec<(String, AgentBehavior)>> {
-    let escaped_agent_did = escape_graphql_string(agent_did);
-    let query = format!(
-        r#"{{
-            AgentBehavior(
-                filter: {{ agent_did: {{ _eq: "{escaped_agent_did}" }} }},
-                order: {{ created_at: ASC }}
-            ) {{
-                _docID
-                behavior_id
-                agent_did
-                display_name
-                system_prompt
-                backend_id
-                model_name
-                tool_selection_id
-                inference_profile_id
-                compaction_strategy
-                compaction_threshold
-                enabled
-                created_at
-            }}
-        }}"#
-    );
-
-    let resp = node.execute(&query).await;
-    if resp.has_errors() {
-        anyhow::bail!("list AgentBehavior failed: {:?}", resp.errors);
-    }
-
-    Ok(serde_helpers::rows_with_doc_id(resp.data.as_ref(), "AgentBehavior"))
 }
 
 pub async fn load_inference_profile(
@@ -542,87 +405,6 @@ pub(crate) async fn list_inference_profile_records(
     Ok(serde_helpers::rows_with_doc_id(resp.data.as_ref(), "InferenceProfile"))
 }
 
-pub async fn upsert_agent_behavior(node: &EmbeddedNode, behavior: &AgentBehavior) -> Result<()> {
-    let escaped_behavior_id = escape_graphql_string(&behavior.behavior_id);
-    let escaped_agent_did = escape_graphql_string(&behavior.agent_did);
-    let created_at = behavior
-        .created_at
-        .as_deref()
-        .map(ToOwned::to_owned)
-        .unwrap_or_else(|| chrono::Utc::now().to_rfc3339());
-
-    let add_fields = vec![
-        Some(format!(r#"behavior_id: "{escaped_behavior_id}""#)),
-        Some(format!(r#"agent_did: "{escaped_agent_did}""#)),
-        graphql_fields::graphql_string_field("display_name", behavior.display_name.as_deref()),
-        graphql_fields::graphql_string_field("system_prompt", behavior.system_prompt.as_deref()),
-        graphql_fields::graphql_string_field("backend_id", behavior.backend_id.as_deref()),
-        graphql_fields::graphql_string_field("model_name", behavior.model_name.as_deref()),
-        graphql_fields::graphql_string_field("tool_selection_id", behavior.tool_selection_id.as_deref()),
-        graphql_fields::graphql_string_field(
-            "inference_profile_id",
-            behavior.inference_profile_id.as_deref(),
-        ),
-        graphql_fields::graphql_string_field(
-            "compaction_strategy",
-            behavior.compaction_strategy.as_deref(),
-        ),
-        graphql_fields::graphql_optional_float_field("compaction_threshold", behavior.compaction_threshold),
-        Some(format!("enabled: {}", graphql_fields::graphql_bool(behavior.enabled))),
-        Some(format!(
-            r#"created_at: "{}""#,
-            escape_graphql_string(created_at.as_str())
-        )),
-    ]
-    .into_iter()
-    .flatten()
-    .collect::<Vec<_>>()
-    .join(",\n                    ");
-
-    let update_fields = vec![
-        Some(format!(r#"agent_did: "{escaped_agent_did}""#)),
-        graphql_fields::graphql_string_field("display_name", behavior.display_name.as_deref()),
-        graphql_fields::graphql_string_field("system_prompt", behavior.system_prompt.as_deref()),
-        graphql_fields::graphql_string_field("backend_id", behavior.backend_id.as_deref()),
-        graphql_fields::graphql_string_field("model_name", behavior.model_name.as_deref()),
-        graphql_fields::graphql_string_field("tool_selection_id", behavior.tool_selection_id.as_deref()),
-        graphql_fields::graphql_string_field(
-            "inference_profile_id",
-            behavior.inference_profile_id.as_deref(),
-        ),
-        graphql_fields::graphql_string_field(
-            "compaction_strategy",
-            behavior.compaction_strategy.as_deref(),
-        ),
-        graphql_fields::graphql_optional_float_field("compaction_threshold", behavior.compaction_threshold),
-        Some(format!("enabled: {}", graphql_fields::graphql_bool(behavior.enabled))),
-    ]
-    .into_iter()
-    .flatten()
-    .collect::<Vec<_>>()
-    .join(",\n                    ");
-
-    let mutation = format!(
-        r#"mutation {{
-            upsert_AgentBehavior(
-                filter: {{ behavior_id: {{ _eq: "{escaped_behavior_id}" }} }},
-                add: {{
-                    {add_fields}
-                }},
-                update: {{
-                    {update_fields}
-                }}
-            ) {{ _docID }}
-        }}"#
-    );
-
-    let resp = node.execute(&mutation).await;
-    if resp.has_errors() {
-        anyhow::bail!("upsert AgentBehavior failed: {:?}", resp.errors);
-    }
-    Ok(())
-}
-
 pub async fn upsert_tool_selection(
     node: &EmbeddedNode,
     selection: &ToolSelectionDocument,
@@ -746,31 +528,6 @@ pub async fn upsert_inference_profile(
         anyhow::bail!("upsert InferenceProfile failed: {:?}", resp.errors);
     }
     Ok(())
-}
-
-async fn create_default_behavior(
-    node: &EmbeddedNode,
-    agent_did: &str,
-    behavior_id: &str,
-) -> Result<()> {
-    upsert_agent_behavior(
-        node,
-        &AgentBehavior {
-            behavior_id: behavior_id.to_string(),
-            agent_did: agent_did.to_string(),
-            display_name: Some(DEFAULT_BEHAVIOR_LABEL.to_string()),
-            system_prompt: None,
-            backend_id: None,
-            model_name: None,
-            tool_selection_id: None,
-            inference_profile_id: None,
-            compaction_strategy: None,
-            compaction_threshold: None,
-            enabled: true,
-            created_at: Some(chrono::Utc::now().to_rfc3339()),
-        },
-    )
-    .await
 }
 
 #[cfg(test)]
